app/backend/requirements.txt
fastapi
uvicorn
pytest
pytest-cov
httpx
sqlalchemy
psycopg2-binary
python-dotenv
pyjwt
pwdlib[argon2]
python-multipart
pydantic[email]